requirements.txt
langchain==0.1.12
openai>=1.0.0
pypdf2==3.0.1
python-dotenv>=0.19.0
streamlit==1.32.2
chromadb>=0.4.0
tiktoken==0.6.0
python-magic==0.4.27
pdfplumber==0.10.4
pandas==2.2.1
numpy==1.26.4
pillow==10.2.0
matplotlib==3.8.3
plotly==5.19.0
pdf2image==1.17.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80 